open_sora/requirements.txt
torch>=2.0.0
transformers>=4.35.0
diffusers>=0.21.0
accelerate>=0.24.0
einops>=0.7.0
xformers>=0.0.22
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0


